+++ public/exercise.html
@@ -5,74 +5,77 @@
 	<meta charset="UTF-8">
 	<title>Evolve Main</title>
 	<!-- Latest compiled and minified CSS & JS -->
-	<link rel="stylesheet" href="../css/style.css" media="screen" title="no title">
+	<link rel="stylesheet" href="./css/style.css" media="screen" title="no title">
 	<link href="https://maxcdn.bootstrapcdn.com/bootswatch/4.0.0/flatly/bootstrap.min.css" rel="stylesheet">
 	<script src="https://code.jquery.com/jquery.js"></script>
 	<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js"></script>
 </head>
 
 <nav class="navbar navbar-dark bg-primary">
-		<a class="navbar-brand" href="#">Evolve</a>
-		<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarText" aria-controls="navbarText" aria-expanded="false" aria-label="Toggle navigation">
-		  <span class="navbar-toggler-icon"></span>
-		</button>
-		<div class="collapse navbar-collapse" id="navbarText">
-		  <ul class="navbar-nav mr-auto">
+	<a class="navbar-brand" href="#">Evolve</a>
+	<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarText"
+		aria-controls="navbarText" aria-expanded="false" aria-label="Toggle navigation">
+		<span class="navbar-toggler-icon"></span>
+	</button>
+	<div class="collapse navbar-collapse" id="navbarText">
+		<ul class="navbar-nav mr-auto">
 			<li class="nav-item active">
-			  <a class="nav-link" href="#">Go To Journal <span class="sr-only">(current)</span></a>
+				<a class="nav-link" href="journal.html">Go To Journal</a>
 			</li>
 			<li class="nav-item">
-			  <a class="nav-link" href="#">Start New Entry</a>
+				<a class="nav-link" href="entries.html">Start New Entry</a>
 			</li>
-		  </ul>
-		</div>
-  </nav>
+		</ul>
+	</div>
+</nav>
+
 <body>
-		 <div class="container-fluid">
-			 <div class="row">
-				 <div class="col-md-8 offset-md-2">
-					<img id="mainPhoto" src="StairsLogo.png" class="rounded mx-auto d-block" alt="Responsive image">
-				 </div>
-			 </div>
-			 <div class="row">
-				<div class="col-md-8 offset-md-2">
-					<div class="author-container">
-						<table class="table table-striped">
-							<thead>
-								<tr>
-									<th>Exercise</th>
-									<th># of Entries</th>
-									<th>See Entries</th>
-									<th>New Entry</th>
-									<th>Delete **</th>
-								</tr>
-							</thead>
-							<tbody>
-								<tr id="form-row">
-									<form id="author-form">
-										<td colspan="4">
-											<input placeholder="Enter a New Exercise" id="author-name" type="text" />
-										</td>
-										<td><button type="submit" class="btn btn-dark">Submit Entry</button>
-										</td>
-									</form>
-								</tr>
-							</tbody>
-						</table>
-					</div>
+	<div class="container-fluid">
+		<div class="row">
+			<div class="col-md-8 offset-md-2">
+				<img id="mainPhoto" src="./css/StairsLogo.png" class="rounded mx-auto d-block" alt="Responsive image">
+			</div>
+		</div>
+		<div class="row">
+			<div class="col-md-8 offset-md-2">
+				<div class="exercise-container">
+					<table class="table table-striped">
+						<thead>
+							<tr>
+								<th>Exercise</th>
+								<th># of Entries</th>
+								<th>See Entries</th>
+								<th>New Entry</th>
+								<th>Delete</th>
+							</tr>
+						</thead>
+						<tbody>
+							<tr id="form-row">
+								<form id="exercise-form">
+									<td colspan="4">
+										<input placeholder="Enter a New Exercise" id="exercise-name" type="text" />
+									</td>
+									<td><button type="submit" class="btn btn-dark">Submit Entry</button>
+									</td>
+								</form>
+							</tr>
+						</tbody>
+					</table>
 				</div>
 			</div>
-		 </div>
-		<!-- Custom Script -->
-		<script src="js/author.js" type="text/javascript"></script>
+		</div>
+	</div>
+	<!-- Custom Script -->
+	<script src="js/exercise.js" type="text/javascript"></script>
 
 </body>
 <footer>
-    <div class="container">
-			<div class="col-12">
-                <p class="text-center">Copyright &copy;2021 by Evolve </p>
-            </div>
-        </div>
-    </div>
+	<div class="container">
+		<div class="col-12">
+			<p class="text-center">Copyright &copy;2021 by Evolve </p>
+		</div>
+	</div>
+	</div>
 </footer>
+
 </html>
