+++ public/journal.html
@@ -12,42 +12,41 @@
 </head>
 
 <nav class="navbar navbar-dark bg-primary">
-	<a class="navbar-brand" href="#">Evolve</a>
-	<button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarText" aria-controls="navbarText" aria-expanded="false" aria-label="Toggle navigation">
-	  <span class="navbar-toggler-icon"></span>
-	</button>
-	<div class="collapse navbar-collapse" id="navbarText">
-	  <ul class="navbar-nav mr-auto">
-		<li class="nav-item active">
-		  <a class="nav-link" href="#">Go To Journal <span class="sr-only">(current)</span></a>
-		</li>
-		<li class="nav-item">
-		  <a class="nav-link" href="#">Start New Entry</a>
-		</li>
-	  </ul>
-	</div>
+  <a class="navbar-brand" href="#">Evolve</a>
+  <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarText"
+    aria-controls="navbarText" aria-expanded="false" aria-label="Toggle navigation">
+    <span class="navbar-toggler-icon"></span>
+  </button>
+  <div class="collapse navbar-collapse" id="navbarText">
+    <ul class="navbar-nav mr-auto">
+      <li class="nav-item active">
+        <a class="nav-link" href="#">Go To Journal <span class="sr-only">(current)</span></a>
+      </li>
+      <li class="nav-item">
+        <a class="nav-link" href="#">Start New Entry</a>
+      </li>
+    </ul>
+  </div>
 </nav>
 
-  <body>
-    <div class="container blog">
-      <div class="row">
-        <div class="col-md-2 col-sm-2">
-          <a class="btn btn-light" href="/cms">New Post</a>
-        </div>
-        <div class="col-md-2 col-sm-2">
-          <a class="btn btn-light" href="/authors">Manage Authors</a>
-        </div>
+<body>
+  <div class="container blog">
+    <div class="row">
+      <div class="col-md-2 col-sm-2">
+        <a class="btn btn-light" href="./entries.html">New Post</a>
       </div>
-      <div class="row">
-        <div class="col-md-12 blog-container">
-        </div>
+      <div class="col-md-2 col-sm-2">
+        <a class="btn btn-light" href="./exercise.html">Manage Authors</a>
       </div>
     </div>
+    <div class="row">
+      <div class="col-md-12 entries-container">
+      </div>
+    </div>
+  </div>
   <!-- Custom Script -->
-  <script src="js/blog.js" type="text/javascript"></script>
+  <script src="js/journal.js" type="text/javascript"></script>
 
 </body>
 
 </html>
-
-
