+++ public/js/entries.js
@@ -1,137 +1,134 @@
 $(document).ready(function () {
-    // entriesContainer holds all of our posts
-    var entriesContainer = $(".entries-container");
-    var postCategorySelect = $("#category");
-    // Click events for the edit and delete buttons
-    $(document).on("click", "button.delete", handlePostDelete);
-    $(document).on("click", "button.edit", handlePostEdit);
-    // Variable to hold our posts
-    var posts;
-
-    // The code below handles the case where we want to get entries for a specific exercise
-    // Looks for a query param in the url for exercise_id
+    // Getting jQuery references to the post body, title, form, and exercise select
+    var bodyInput = $("#body");
+    var titleInput = $("#title");
+    var entriesForm = $("#entries");
+    var exerciseSelect = $("#exercise");
+    // Adding an event listener for when the form is submitted
+    $(entriesForm).on("submit", handleFormSubmit);
+    // Gets the part of the url that comes after the "?" (which we have if we're updating a post)
     var url = window.location.search;
+    var postId;
     var exerciseId;
-    if (url.indexOf("?exercise_id=") !== -1) {
-        exerciseId = url.split("=")[1];
-        getPosts(exerciseId);
+    // Sets a flag for whether or not we're updating a post to be false initially
+    var updating = false;
+
+    // If we have this section in our url, we pull out the post id from the url
+    // In '?post_id=1', postId is 1
+    if (url.indexOf("?post_id=") !== -1) {
+        postId = url.split("=")[1];
+        getPostData(postId, "post");
     }
-    // If there's no exerciseId we just get all entries as usual
-    else {
-        getPosts();
+    // Otherwise if we have an exercise_id in our url, preset the exercise select box to be our Exercise
+    else if (url.indexOf("?exercise_id=") !== -1) {
+        exerciseId = url.split("=")[1];
     }
 
-    // This function grabs posts from the database and updates the view
-    function getPosts(exercise) {
-        exerciseId = exercise || "";
-        if (exerciseId) {
-            exerciseId = "/?exercise_id=" + exerciseId;
+    // Getting the exercises, and their related posts
+    getExercises();
+
+    // A function for handling what happens when the form to create a new post is submitted
+    function handleFormSubmit(event) {
+        event.preventDefault();
+        // Wont submit the post if we are missing a body, title, or exercise
+        if (!titleInput.val().trim() || !bodyInput.val().trim() || !exerciseSelect.val()) {
+            return;
         }
-        $.get("/api/posts" + exerciseId, function (data) {
-            console.log("Posts", data);
-            posts = data;
-            if (!posts || !posts.length) {
-                displayEmpty(exercise);
-            }
-            else {
-                initializeRows();
+        // Constructing a newPost object to hand to the database
+        var newPost = {
+            title: titleInput
+                .val()
+                .trim(),
+            body: bodyInput
+                .val()
+                .trim(),
+            ExerciseId: exerciseSelect.val()
+        };
+
+        // If we're updating a post run updatePost to update a post
+        // Otherwise run submitPost to create a whole new post
+        if (updating) {
+            newPost.id = postId;
+            updatePost(newPost);
+        }
+        else {
+            submitPost(newPost);
+        }
+    }
+
+    // Submits a new post and brings user to blog page upon completion
+    function submitPost(post) {
+        $.post("/api/posts", post, function () {
+            window.location.href = "/entries";
+        });
+    }
+
+    // Gets post data for the current post if we're editing, or if we're adding to an exercise's existing posts
+    function getPostData(id, type) {
+        var queryUrl;
+        switch (type) {
+            case "post":
+                queryUrl = "/api/posts/" + id;
+                break;
+            case "exercise":
+                queryUrl = "/api/exercises/" + id;
+                break;
+            default:
+                return;
+        }
+        $.get(queryUrl, function (data) {
+            if (data) {
+                console.log(data.ExerciseId || data.id);
+                // If this post exists, prefill our entries forms with its data
+                titleInput.val(data.title);
+                bodyInput.val(data.body);
+                exerciseId = data.ExerciseId || data.id;
+                // If we have a post with this id, set a flag for us to know to update the post
+                // when we hit submit
+                updating = true;
             }
         });
     }
 
-    // This function does an API call to delete posts
-    function deletePost(id) {
+    // A function to get Exercises and then render our list of Exercises
+    function getExercises() {
+        $.get("/api/exercises", renderExerciseList);
+    }
+    // Function to either render a list of exercises, or if there are none, direct the user to the page
+    // to create an exercise first
+    function renderExerciseList(data) {
+        if (!data.length) {
+            window.location.href = "/exercises";
+        }
+        $(".hidden").removeClass("hidden");
+        var rowsToAdd = [];
+        for (var i = 0; i < data.length; i++) {
+            rowsToAdd.push(createExerciseRow(data[i]));
+        }
+        exerciseSelect.empty();
+        console.log(rowsToAdd);
+        console.log(exerciseSelect);
+        exerciseSelect.append(rowsToAdd);
+        exerciseSelect.val(exerciseId);
+    }
+
+    // Creates the exercise options in the dropdown
+    function createExerciseRow(exercise) {
+        var listOption = $("<option>");
+        listOption.attr("value", exercise.id);
+        listOption.text(exercise.name);
+        return listOption;
+    }
+
+    // Update a given post, bring user to the entries page when done
+    function updatePost(post) {
         $.ajax({
-            method: "DELETE",
-            url: "/api/posts/" + id
+            method: "PUT",
+            url: "/api/posts",
+            data: post
         })
             .then(function () {
-                getPosts(postCategorySelect.val());
+                window.location.href = "/entries";
             });
     }
-
-    // InitializeRows handles appending all of our constructed entries HTML inside entriesContainer
-    function initializeRows() {
-        entriesContainer.empty();
-        var postsToAdd = [];
-        for (var i = 0; i < posts.length; i++) {
-            postsToAdd.push(createNewRow(posts[i]));
-        }
-        entriesContainer.append(postsToAdd);
-    }
-
-    // This function constructs an entry's HTML
-    function createNewRow(post) {
-        var formattedDate = new Date(post.createdAt).toLocaleDateString();
-        var newPostCard = $("<div>");
-        newPostCard.addClass("card");
-        var newPostCardHeading = $("<div>");
-        newPostCardHeading.addClass("card-header");
-        var deleteBtn = $("<button>");
-        deleteBtn.text("x");
-        deleteBtn.addClass("delete btn btn-danger");
-        var editBtn = $("<button>");
-        editBtn.text("EDIT");
-        editBtn.addClass("edit btn btn-info");
-        var newPostTitle = $("<h2>");
-        var newPostDate = $("<small>");
-        var newPostExercise = $("<h5>");
-        newPostExercise.text("Exercise Category: " + post.Exercise.name);
-        newPostAuthor.css({
-            float: "right",
-            color: "blue",
-            "margin-top":
-                "-10px"
-        });
-        var newPostCardBody = $("<div>");
-        newPostCardBody.addClass("card-body");
-        var newPostBody = $("<p>");
-        newPostTitle.text(post.title + " ");
-        newPostBody.text(post.body);
-        newPostDate.text(formattedDate);
-        newPostTitle.append(newPostDate);
-        newPostCardHeading.append(deleteBtn);
-        newPostCardHeading.append(editBtn);
-        newPostCardHeading.append(newPostTitle);
-        newPostCardHeading.append(newPostAuthor);
-        newPostCardBody.append(newPostBody);
-        newPostCard.append(newPostCardHeading);
-        newPostCard.append(newPostCardBody);
-        newPostCard.data("post", post);
-        return newPostCard;
-    }
-
-    // This function figures out which entry we want to delete and then calls deletePost
-    function handlePostDelete() {
-        var currentPost = $(this)
-            .parent()
-            .parent()
-            .data("post");
-        deletePost(currentPost.id);
-    }
-
-    // This function figures out which post we want to edit and takes it to the appropriate url
-    function handlePostEdit() {
-        var currentPost = $(this)
-            .parent()
-            .parent()
-            .data("post");
-        window.location.href = "/journal?post_id=" + currentPost.id;
-    }
-
-    // This function displays a message when there are no posts
-    function displayEmpty(id) {
-        var query = window.location.search;
-        var partial = "";
-        if (id) {
-            partial = " for Exercise #" + id;
-        }
-        entriesContainer.empty();
-        var messageH2 = $("<h2>");
-        messageH2.css({ "text-align": "center", "margin-top": "50px" });
-        messageH2.html("No entries yet" + partial + ", navigate <a href='/journal" + query +
-            "'>here</a> in order to get started.");
-        entriesContainer.append(messageH2);
-    }
-
 });
