+++ server.js
@@ -24,6 +24,7 @@ app.use(passport.session());
 require("./routes/html-routes.js")(app);
 require("./routes/exercise-api-routes.js")(app);
 require("./routes/post-api-routes.js")(app);
+require("./routes/api-routes.js")(app);
 
 // Syncing our database and logging a message to the user upon success
 db.sequelize.sync().then(() => {
